Add component setup
