+++ src/components/ExchangeRates/ExchangeRates.js
@@ -1,0 +1,30 @@
+import {
+  useQuery,
+  gql
+} from "@apollo/client";
+
+const ExchangeRates = () => {
+  const { loading, error, data } = useQuery(EXCHANGE_RATES);
+
+  if (loading) return <p>Loading...</p>;
+  if (error) return <p>Error :(</p>;
+
+  return data.rates.map(({ currency, rate }) => (
+    <div key={currency}>
+      <p>
+        {currency}: {rate}
+      </p>
+    </div>
+  ));
+}
+
+const EXCHANGE_RATES = gql`
+  query GetExchangeRates {
+    rates(currency: "USD") {
+      currency
+      rate
+    }
+  }
+`;
+
+export default ExchangeRates
